hw7

+++ scrapper/src/main/java/edu/java/controller/TgChatApiController.java
@@ -1,12 +1,11 @@
 package edu.java.controller;
 
-import com.fasterxml.jackson.databind.ObjectMapper;
 import edu.java.model.dto.Chat;
 import edu.java.service.ChatService;
+import io.github.bucket4j.Bucket;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.enums.ParameterIn;
 import io.swagger.v3.oas.annotations.media.Schema;
-import jakarta.servlet.http.HttpServletRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,19 +18,14 @@ import org.springframework.web.bind.annotation.RestController;
 public class TgChatApiController implements TgChatApi {
     @Autowired
     private final ChatService chatService;
-
     private static final Logger LOGGER = LoggerFactory.getLogger(TgChatApiController.class);
 
-    private final ObjectMapper objectMapper;
-
-    private final HttpServletRequest request;
-    private final String acceptString = "Accept";
+    @Autowired
+    private Bucket bucket;
 
     @Autowired
-    public TgChatApiController(ChatService chatService, ObjectMapper objectMapper, HttpServletRequest request) {
+    public TgChatApiController(ChatService chatService) {
         this.chatService = chatService;
-        this.objectMapper = objectMapper;
-        this.request = request;
     }
 
     public ResponseEntity<Void> tgChatIdDelete(
@@ -41,19 +35,25 @@ public class TgChatApiController implements TgChatApi {
                    schema = @Schema())
         @PathVariable("id") Long id
     ) {
-        chatService.removeChat(id);
-        return new ResponseEntity<Void>(HttpStatus.OK);
+        if (bucket.tryConsume(1)) {
+            chatService.removeChat(id);
+            return new ResponseEntity<Void>(HttpStatus.OK);
+        } else {
+            return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).build();
+        }
     }
 
     public ResponseEntity<Void> tgChatIdPost(
         @Parameter(in = ParameterIn.PATH, description = "", required = true, schema = @Schema()) @PathVariable("id")
         Long id
     ) {
-        Chat chat = new Chat();
-        chat.setChatId(id);
-
-        chatService.addChat(chat);
-        return new ResponseEntity<Void>(HttpStatus.OK);
+        if (bucket.tryConsume(1)) {
+            Chat chat = new Chat();
+            chat.setChatId(id);
+            chatService.addChat(chat);
+            return new ResponseEntity<Void>(HttpStatus.OK);
+        } else {
+            return ResponseEntity.status(HttpStatus.TOO_MANY_REQUESTS).build();
+        }
     }
-
 }
